Rename dart files according to the class name
