+++ lib/src/widgets/slider_widget.dart
