+++ lib/src/widgets/text_alignment_widget.dart
@@ -29,7 +29,7 @@
 /// Authors: Tony Chen
 
 import 'package:flutter/material.dart';
-import 'package:markdown_widgets/src/utils/text_utils.dart';
+import 'package:markdown_widgets/src/utils/justify_text.dart';
 import 'package:markdown_widgets/src/constants/pkg.dart'
     show contentWidthFactor, screenWidth;
